Makes it so that no warning is issue
After long hours of reading, and trying things out and a couple more of instruction step debugging, I figured out how to stop that warning on qtserialport old qt4 branch version. It has less to do with threads than we expect, in fact it's complaining about a pair of child-parent QObjects being in different threads. Setting a null parent on that case removes the warning. That issue comes from QSocketNotifier  constructor in qtserialport's case. In this case it's QNetworkAccessManager creation that issues it.
Thanks to this great simple example of yours which was a lot easier to fix than that backport library which has a similar "quirk"! I was really close to giving up until I found and fixed your example, which boosted me with hope!
Have you got that already?
Cheers!